test: merge serial tests in bindinfo, expression, parser and statistics
