+++ bindinfo/bind_test.go
