+++ bindinfo/capture_test.go
