+++ bindinfo/handle_test.go
