+++ bindinfo/optimize_test.go
