+++ statistics/handle/ddl_test.go
